requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
multidict==4.7.6
oauthlib==3.1.0
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
slackclient==2.7.2
tweepy==3.8.0
urllib3==1.25.9
yarl==1.4.2
